requirements.txt
flask
transformers
python-dotenv==1.0.0
openai==1.0.0